backend/requirements.txt
# ─────────────────────────────────────────────────────────────────────────
# SplitLedger — Production Dependencies
# Python 3.11+
# Install: pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────────────────

# ── Web framework ─────────────────────────────────────────────────────────
Flask==3.0.3
Werkzeug==3.0.3

# ── Database ──────────────────────────────────────────────────────────────
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
flask-marshmallow==1.1.0
marshmallow-sqlalchemy==1.4.2
psycopg2-binary==2.9.9          # PostgreSQL adapter (binary = no build tools needed)
alembic==1.13.3                 # Database migrations

# ── Validation ────────────────────────────────────────────────────────────
marshmallow==3.22.0             # Schema validation

# ── Authentication ────────────────────────────────────────────────────────
PyJWT==2.9.0                    # JWT access tokens
bcrypt==4.2.0                   # Password hashing (cost factor 12)

# ── Utilities ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1            # Load .env into os.environ
